requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.2
icalendar==5.0.14